feat: 메세지 모달에도 글꼴 적용
