+++ src/components/RecipientPage/MessageCardList/MessageCard/MessageCard.js
@@ -5,11 +5,9 @@ import MessageCardModal from "../../MessageCardModal/MessageCardModal";
 import MessageCardContainer from "../MessageCardContainer/MessageCardContainer";
 import CreatedAt from "./CreatedAt/CreatedAt";
 import SenderInfo from "./SenderInfo/SenderInfo";
+import MessageCardContent from "./MessageCardContent/MessageCardContent";
 
 const MessageCard = ({ message, isEditing, onDelete }) => {
-  const CONTENT_ELEMENT = useRef();
-  const FONT = message.font.replaceAll(" ", "");
-
   const [messageCardModalOpen, setMessageCardModalOpen] = useState(false);
 
   const handleClick = () => {
@@ -18,20 +16,11 @@ const MessageCard = ({ message, isEditing, onDelete }) => {
     }
   };
 
-  useEffect(() => {
-    const CONTENT = message.content;
-    CONTENT_ELEMENT.current.innerHTML = CONTENT;
-  }, []);
-
   return (
     <MessageCardContainer onClick={handleClick}>
       <SenderInfo message={message} isEditing={isEditing} onDelete={onDelete} />
       <div className={styles.DividingLine}></div>
-      <div
-        ref={CONTENT_ELEMENT}
-        className={`${styles.CardContent} ${styles[FONT]}`}
-        // dangerouslySetInnerHTML={{ __html: message.content }}
-      ></div>
+      <MessageCardContent message={message} />
       <CreatedAt createdAt={message.createdAt} />
       {messageCardModalOpen && (
         <MessageCardModal
